# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Goyim Arena API",
    "version": "0.1.0",
    "description": "Public HTTP contract of Goyim Arena. Versioned per the master plan: routes registered by the binary must match this document; the contract test fails on drift.",
    "summary": "Arena API contract"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Versioned API surface served by the arena binary."
    }
  ],
  "paths": {
...
    "/api/v1/me/billing/checkout": {
      "post": {
        "operationId": "createBillingCheckout",
        "summary": "Create a server-priced checkout",
        "description": "Creates one server-authoritative checkout for the owner's account and returns the redirect URL with the priced amount. Amount, currency and return URLs come from the versioned catalog and allowlisted configuration; provider identifiers never serialize. Served with Cache-Control: private, no-store (THR-CACHE-01).",
        "tags": ["billing"],
        "security": [{ "SessionCookie": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BillingCheckoutRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Checkout created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BillingCheckout" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
    "/api/v1/me/billing/subscription": {
      "get": {
        "operationId": "getBillingSubscription",
        "summary": "Member subscription projection",
        "description": "Returns the owner's Member subscription projection from the local mirror: lifecycle and period only. Provider identifiers never serialize. Served with Cache-Control: private, no-store (THR-CACHE-01).",
        "tags": ["billing"],
        "security": [{ "SessionCookie": [] }],
        "responses": {
          "200": {
            "description": "Subscription projection.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BillingSubscription" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/api/v1/me/billing/portal": {
      "post": {
        "operationId": "createBillingPortal",
        "summary": "Open the customer portal",
        "description": "Opens the hosted customer portal for the owner's stored customer and returns only the portal URL. Provider identifiers never serialize. Served with Cache-Control: private, no-store (THR-CACHE-01).",
        "tags": ["billing"],
        "security": [{ "SessionCookie": [] }],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BillingPortalRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Portal session.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BillingPortal" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Validation" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BillingCheckoutRequest": {
        "type": "object",
        "description": "Minimal purchase intent the browser may name. Amount, currency, price and return URLs are resolved by the server.",
        "required": ["market", "product", "idempotency_key"],
        "properties": {
          "market": { "type": "string", "enum": ["BR", "INTERNATIONAL"] },
          "product": { "type": "string" },
          "idempotency_key": { "type": "string" }
        }
      },
      "BillingCheckout": {
        "type": "object",
        "description": "Private checkout document: local intent, priced money and redirect URL. No provider identifier serializes.",
        "required": ["intent_id", "status", "redirect_url", "amount_minor", "currency", "market", "product", "replayed"],
        "properties": {
          "intent_id": { "type": "string" },
          "status": { "type": "string", "enum": ["created", "open", "paid", "expired", "failed"] },
          "redirect_url": { "type": "string" },
          "amount_minor": { "type": "integer", "minimum": 1, "description": "Exact price in minor units of the currency." },
          "currency": { "type": "string", "enum": ["BRL", "USD"] },
          "market": { "type": "string", "enum": ["BR", "INTERNATIONAL"] },
          "product": { "type": "string" },
          "replayed": { "type": "boolean" }
        }
      },
      "BillingSubscription": {
        "type": "object",
        "description": "Private Member subscription projection: lifecycle and period only. No provider identifier serializes.",
        "required": ["has_subscription"],
        "properties": {
          "has_subscription": { "type": "boolean" },
          "status": { "type": "string", "enum": ["incomplete", "incomplete_expired", "trialing", "active", "past_due", "canceled", "unpaid", "paused"] },
          "product": { "type": "string" },
          "market": { "type": "string", "enum": ["BR", "INTERNATIONAL"] },
          "current_period_end": { "type": ["string", "null"], "format": "date-time" },
          "cancel_at_period_end": { "type": "boolean" }
        }
      },
      "BillingPortalRequest": {
        "type": "object",
        "description": "Optional operation token for an idempotent portal opening.",
        "properties": {
          "idempotency_key": { "type": "string" }
        }
      },
      "BillingPortal": {
        "type": "object",
        "description": "Hosted portal session. Only the portal URL leaves the billing module.",
        "required": ["portal_url"],
        "properties": {
          "portal_url": { "type": "string" }
        }
      },
...
      "Problem": {
        "description": "RFC 9457 Problem Details document returned by every error response, with media type application/problem+json.",
        "type": "object",
        "required": ["type", "title", "status", "code"],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference",
            "description": "Stable URI identifying the problem kind.",
            "examples": ["https://goyim.dev/problems/validation"]
          },
          "title": {
            "type": "string",
            "description": "Short human-readable summary, generic for internal errors."
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code, identical to the response status."
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code, e.g. ARENA-validation."
          },
          "request_id": {
            "type": "string",
            "description": "Correlation ID echoed by the X-Request-Id middleware; join it with server logs."
          },
          "detail": {
            "type": "string",
            "description": "Public safe detail. Suppressed for internal errors: causes never serialize to clients."
          }
        }
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "arena_session",
        "description": "Opaque session token cookie (Secure, HttpOnly, SameSite). Sessions arrive with the identity phase; the scheme is declared now to anchor the contract."
      },
      "CsrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Per-request CSRF token required by state-changing requests once sessions exist."
      }
    },
...
    "responses": {
      "Validation": {
        "description": "Request validation failed.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication is required.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "Forbidden": {
        "description": "The authenticated caller is not allowed to do this.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
...
    },
...
  },
...
  "security": [],
...
}